✨[Feat] done modify and remove review
